requirements.txt
python-telegram-bot>=20.0
SQLAlchemy>=2.0
python-dotenv>=1.0
requests>=2.31